+++ NEWS.md
@@ -1,13 +1,20 @@
 Release notes
 =============
 
+Version 0.3.1 (2023-02-16)
+--------------------------
+### Introduction of linepacking
+* Redefinition of `PipelineMode` as abstract type `PipeMode` and introduction of `PipeSimple` as a composite type corresponding to the previous `PipelineMode`
+* Introduction of a simple linepacking implementation _via_ the type `PipeLinepackSimple
+* Change of `Area` to `abstract type` to be able to dispatch on areas
+* Rewriting how functions for variable generation are called for easier introduction of variables for different `TransmissionMode`s
+
 Version 0.3.0 (2023-02-02)
 --------------------------
 ### Adjustmends to updates in EnergyModelsBase
 Adjustment to version 0.3.0, namely:
 * The removal of emissions from `Node` type definition that do not require them in all tests
 * Removal of the type `GlobalData` and replacement with fields in the type `OperationalModel` in all tests
-
 Version 0.2.2 (2022-12-12)
 --------------------------
 ### Internal release